feat(surface): let dsurface report the atom under the cursor
Clicking or hovering a direct molecular surface did nothing: dsurface
never took part in hit testing. It already knows the answer, though --
every mesh vertex carries the id of the atom it belongs to, cached
alongside the geometry -- and the GPU pick pass is built to read exactly
that.

So the renderer only has to say so. buildGpuMesh() writes each vertex's
owning atom into the hit-name attribute the triangle primitive already
has, and displayPick() reuses display(), whose primitive switches to its
pick program on its own. The result comes back through MolRenderer's
existing atom-level interpretation, so the hover chip, the click label
and the hover highlight all work without a change on the tritium side;
the highlight covers that atom's patch, which is the unit a surface
makes sense in.

Only the fill draw mode is pickable. Wireframe and dots go through the
display list, which carries one name for a whole mesh, and a surface
without a shader is not drawn into the ID buffer at all rather than
occluding what is behind it under no name. A vertex whose owning atom is
unknown encodes to "no name" and is simply not picked, and vertices
hidden by showsel are never uploaded, so neither needs its own gate.

Rectangle and lasso selection still ignore the surface: those run the CPU
point test, where a surface could only offer its atoms' centres, which is
not where the patch the user clicked is.
